test(OAuthExt): fix test build and update to use v3 xUnit and implement xunit.runner.json configured usage
