+++ Yllibed.HttpServer.Handlers.Uno.Tests/xunit.runner.json
@@ -1,0 +1,7 @@
+{
+    "$schema": "https://xunit.net/schema/current/xunit.runner.schema.json",
+    "methodDisplay": "method",
+    "methodDisplayOptions": "replaceUnderscopreWithSpace",
+    "preEnumerateTheories": true,
+    "stopOnFail": true
+}
